fix stale check

+++ primitives/src/lib.rs
@@ -416,5 +416,5 @@ pub mod staking {
 
 pub type Score = Vec<Point>;
 pub type Point = u64;
-pub type Round = u64;
+pub type Round = u32;
 pub type ScoreSignature = u64;
